% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{accuracy}
\alias{accuracy}
\title{A computer of prediction accuracy}
\usage{
accuracy(pred_probs, obs_result)
}
\arguments{
\item{pred_probs}{the prediction probability matrix}

\item{obs_result}{the winner index vector}
}
\description{
A computer of prediction accuracy
}
