% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{actual_profit_sum}
\alias{actual_profit_sum}
\title{A computer of prediction accuracy}
\usage{
actual_profit_sum(pred_probs, odds, obs_outcomes)
}
\arguments{
\item{pred_probs}{the prediction probability matrix}

\item{odds}{the profit odds given by gamble company}

\item{obs_outcomes}{the match outcome matrix}
}
\description{
A computer of prediction accuracy
}
