% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeOut}
\alias{ComputeOut}
\title{A computer of prediction accuracy}
\usage{
ComputeOut(newdata)
}
\arguments{
\item{newdata}{the validation set of match outcomes}
}
\description{
A computer of prediction accuracy
}
