% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalPoissonLogLC.R
\name{data}
\alias{data}
\title{A dataset used for HPL.}
\description{
This dataset is used to fit the HPL model
}
\examples{
\dontrun{
data(data)
attach(data)
b<-HPL(data,size_train = 360,size_valid = 20)
}
}
