% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_cpp}
\alias{gibbs_cpp}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbs_cpp(N, thin, n, a, b)
}
\arguments{
\item{N}{the number of samples}

\item{thin}{the number of between-sample random numbers}

\item{n, a, b}{the shape parameters concerned with binomial and beta distribution}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
gibbs_cpp(100, 10 , 10,5,5)
}
}
