% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalPoissonLogLC.R
\name{HPL}
\alias{HPL}
\title{A hierarchical Poisson Linear model for Premier League.}
\usage{
HPL(data, size_train, size_valid)
}
\arguments{
\item{data}{used data.}

\item{size_train}{A number of training set.}

\item{size_valid}{A number of prediction or validation set.}
}
\value{
The summary of prediction and profit by gambling.
}
\description{
A hierarchical Poisson Linear model for Premier League.
}
\examples{
\dontrun{
data(data)
attach(data)
b<-HPL(data,size_train = 360,size_valid = 20)
}
}
