% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ideal_profit_sum}
\alias{ideal_profit_sum}
\title{A computer of prediction accuracy}
\usage{
ideal_profit_sum(pred_probs, odds)
}
\arguments{
\item{pred_probs}{the prediction probability matrix}

\item{odds}{the profit odds given by gamble company}
}
\description{
A computer of prediction accuracy
}
