% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RPS}
\alias{RPS}
\title{A computer of Ranked Probability Score}
\usage{
RPS(pred_probs, obs_outcomes)
}
\arguments{
\item{pred_probs}{the prediction probability matrix}

\item{obs_outcomes}{the match outcome matrix}
}
\description{
A computer of Ranked Probability Score
}
