% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_max}
\alias{sum_max}
\title{A computer of prediction accuracy}
\usage{
sum_max(pred_probs)
}
\arguments{
\item{pred_probs}{the prediction probability matrix}
}
\description{
A computer of prediction accuracy
}
